This is at least a partial resolution for https://github.com/google/closure-compiler/issues/478.

Specifying the `--no-output` (EDIT: this has been changed to `--checks-only`) command line parameter skips output and runs only checks.
It works in with each of the compilation levels.

---

Alternate approaches I considered:

* **Add a new compilation level** However, unlike GCC's `-O`, Closure's compilation level determines a lot more than just the optimizations. For example, `checkTypes` is enabled by default `ADVANCED_OPTIMIZATIONS`, but not `SIMPLE_OPTIMIZATIONS`, and never for `WHITESPACE_ONLY`.

* **Use `skipNormalOutputs` in `CommandLineConfig`**. This  sounds similar, except it also skips checks. (I believe it is meant for outputting only dependency info.)

---

Performance benefits:

* 6000-line file with `WHITESPACE_ONLY` -- from 2.6 to 1.7 seconds
* 6000-line file with `SIMPLE_OPTIMIZATIONS` -- from 6.1 seconds to 3.8 seconds
* Large project with `ADVANCED_OPTIMIZATIONS` -- from 135 seconds to 70 seconds