@concavelenz 2015-03-31
@MatrixFrog Do you have any outstanding concerns?  I think we should merge this.