@MatrixFrog 2015-03-31
There is the general problem that we have a few different "only" flags and it can get confusing. But this is a good one to have so I'm not too worried about it.