@pauldraper 2015-04-04
I rebased to make sure tests still pass.